+++ octasine/src/gui/mod.rs
@@ -457,7 +457,6 @@ impl<H: GuiSyncHandle> Application for OctaSineIcedApplication<H> {
     #[cfg(feature = "wgpu")]
     fn renderer_settings() -> iced_baseview::renderer::Settings {
         iced_baseview::renderer::Settings {
-            present_mode: iced_baseview::renderer::wgpu::PresentMode::Immediate,
             default_font: Some(OPEN_SANS_BYTES_SEMI_BOLD),
             default_text_size: FONT_SIZE,
             antialiasing: Some(iced_baseview::renderer::settings::Antialiasing::MSAAx4),
@@ -471,6 +470,9 @@ impl<H: GuiSyncHandle> Application for OctaSineIcedApplication<H> {
         iced_baseview::renderer::Settings {
             default_font: Some(OPEN_SANS_BYTES_SEMI_BOLD),
             default_text_size: FONT_SIZE,
+            #[cfg(target_os = "linux")]
+            antialiasing: Some(iced_baseview::renderer::settings::Antialiasing::MSAAx4),
+            #[cfg(not(target_os = "linux"))]
             antialiasing: Some(iced_baseview::renderer::settings::Antialiasing::MSAAx8),
             text_multithreading: false,
         }
@@ -867,6 +869,10 @@ pub fn get_iced_baseview_settings<H: GuiSyncHandle>(
             title: plugin_name,
             #[cfg(feature = "glow")]
             gl_config: Some(iced_baseview::baseview::gl::GlConfig {
+                // 4x MSAA on Linux solves crashes for some people
+                #[cfg(target_os = "linux")]
+                samples: Some(4),
+                #[cfg(not(target_os = "linux"))]
                 samples: Some(8),
                 ..Default::default()
             }),