+++ scripts/macos/build-vst2-and-install.sh
@@ -4,7 +4,7 @@
 
 set -e
 
-cargo xtask bundle octasine --profile "release-debug" --no-default-features --features "glow vst2"
+cargo xtask bundle octasine --profile "release-debug"
 
 TARGET="/Library/Audio/Plug-Ins/VST/OctaSine.vst"